+++ Migrations/WhatsOnTapContextModelSnapshot.cs
@@ -37,7 +37,7 @@ namespace WhatsOnTap.Migrations
 
                     b.Property<double>("srm");
 
-                    b.Property<int>("styleId");
+                    b.Property<long>("styleId");
 
                     b.HasKey("id");
 
@@ -95,7 +95,7 @@ namespace WhatsOnTap.Migrations
                     b.Property<long?>("id")
                         .ValueGeneratedOnAdd();
 
-                    b.Property<int>("beerId");
+                    b.Property<long>("beerId");
 
                     b.Property<bool>("isEmpty");
